README tweaks.
More spelling fixes, also changed "deployment type" to "deployment" in
the sentence describing what a Check is because as I understand it so
far, each deployment may have many alerts/checks/incidents (presuming
these are synonymous so far)? A deployment type says to me that we
record different categories of deployments, but I see nothing other than
instances of deployments so far.