Updated notes and such.
Had problems running the curl to post an alert as-is, fixed that up.
Have added a CHEATSHEET for me to stash some notes for the Atom editor
(not too bad so far, but may be missing some stuff I like) and Go. Still
studying the app and tracing through the code ... I see where some of
the next steps are leading now, methinks.