Notifier work started.
First, the alert handler now takes an array of Listener interfaces and a
notifier instance is now passed to it in addition to the alerter.

I'd considered making model_pagerduty and model_slack simply implement
the listener interface, too, but then couldn't find a decent place for
the given `getInfo` method ... plus deciding to not use goroutines as a
first shot ... even as I type this I realize these are all very
negotiable things, so I'll just run in a direction until it seems
suboptimal.

So with the notifier staying around, I added a Transmitter interface
that the specific models can implement and I have that basic
functionality working in notifier, plus a test.