Notifier now pushes an AlertPackage to Transmitter
This allows each implementer to have full access to all the data to act
on it it whatever way they need. e.g. Slack is the same action
regardless of resolution, but PagerDuty needs separate actions based on
resolution status. This should be by-value if I'm paying attention, to
avoid a malformed transmitter from screwing up the data for other
transmitters. Hmmm, I could test that.

Test setup now allows a test database to be in play. Though I want to
extract that.