Moved hardcoded PagerDuty key to Settings.
Also, skipping the notification if no setting exists. Added a
TransmitResult to better verify the results in the tests and setup
production to track it if it needs to/wants to in some way. Right now
it's just logging error cases, which I think out loud about in
JOURNAL.md.

Need to do a proper integration test next, in case the unit test is
making some improper assumptions about how this all hangs together.