Adding tests in prep of CurrentChecks refactor.
Tests first uncovered a missing index: "Index `type` was not found on
table `alerts_test.checks`". Dunno what of my setup code is leftover
from the original exercise code and what's my screwup, but the code to
create the index wasn't being applied to the proper database -- plus
there was no 'type' index anyway. So, now that that's put to bed, and I
understand things a bit better, onward with the refactoring.